requirements.txt
https://www.djangoproject.com/download/1.7.b4/tarball/
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
celery==3.1.12
django-admin-bootstrapped==1.6.5
django-nose==1.2
djangorestframework==2.3.13
ecdsa==0.11
fuzzywuzzy==0.2
gunicorn==19.0.0
kombu==3.0.19
nose==1.3.3
paramiko==1.14.0
psycopg2==2.5.3
pycrypto==2.6.1
pytz==2014.4
redis==2.10.1
selenium==2.42.1
splinter==0.6.0
sure==1.2.7
beautifulsoup4
behave
behaving
git+git://github.com/eliziario/django-behave.git
